+++ src/Settings.h
@@ -41,8 +41,8 @@ public:
         bool extraRingModeChanged {false};
         bool enchantmentStrengthChanged {false};
         bool fingerSelectionChanged {false};
-        bool bondOfMatrimonyLeftRingFingerChanged {false};
-        bool npcBondOfMatrimonyLeftRingFingerPreferenceChanged {false};
+        bool playerAlwaysEquipBondOfMatrimonyLeftRingFingerChanged {false};
+        bool npcAlwaysEquipBondOfMatrimonyLeftRingFingerChanged {false};
         bool npcSupportChanged {false};
         bool npcSupportEnabled {true};
         bool unequipAllClearsExtraRingsChanged {false};
@@ -53,8 +53,8 @@ public:
             return extraRingModeChanged
                    || enchantmentStrengthChanged
                    || fingerSelectionChanged
-                   || bondOfMatrimonyLeftRingFingerChanged
-                   || npcBondOfMatrimonyLeftRingFingerPreferenceChanged
+                   || playerAlwaysEquipBondOfMatrimonyLeftRingFingerChanged
+                   || npcAlwaysEquipBondOfMatrimonyLeftRingFingerChanged
                    || npcSupportChanged
                    || unequipAllClearsExtraRingsChanged
                    || virtualSlotsChanged;
@@ -71,8 +71,8 @@ public:
     [[nodiscard]] std::uint32_t GetFingerSelectModifierKey() const;
     [[nodiscard]] std::uint32_t GetFingerSelectModifierButton() const;
     [[nodiscard]] bool IsNpcSupportEnabled() const;
-    [[nodiscard]] bool ShouldUseBondOfMatrimonyOnLeftRingFinger() const;
-    [[nodiscard]] bool ShouldNpcPreferBondOfMatrimonyOnLeftRingFinger() const;
+    [[nodiscard]] bool ShouldPlayerAlwaysEquipBondOfMatrimonyOnLeftRingFinger() const;
+    [[nodiscard]] bool ShouldNpcAlwaysEquipBondOfMatrimonyOnLeftRingFinger() const;
     [[nodiscard]] bool IsActorVirtualRingSupportEnabled(Core::ActorKey a_actor) const;
     [[nodiscard]] bool ShouldUnequipAllClearExtraRings() const;
     [[nodiscard]] bool IsTargetEnabled(Core::Target a_target) const;
@@ -87,8 +87,8 @@ private:
     std::atomic<std::uint32_t> fingerSelectModifierKey_ {kDefaultFingerSelectModifierKey};
     std::atomic<std::uint32_t> fingerSelectModifierButton_ {kDefaultFingerSelectModifierButton};
     std::atomic_bool npcSupportEnabled_ {true};
-    std::atomic_bool bondOfMatrimonyOnLeftRingFinger_ {false};
-    std::atomic_bool npcPreferBondOfMatrimonyOnLeftRingFinger_ {true};
+    std::atomic_bool playerAlwaysEquipBondOfMatrimonyOnLeftRingFinger_ {false};
+    std::atomic_bool npcAlwaysEquipBondOfMatrimonyOnLeftRingFinger_ {true};
     std::atomic_bool unequipAllClearsExtraRings_ {true};
     std::atomic<std::uint16_t> enabledVirtualTargetBits_ {kDefaultEnabledVirtualTargetBits};
 };